move listeners to separate package
