+++ tia-client/src/main/resources/META-INF/services/org.junit.platform.launcher.TestExecutionListener
@@ -1,1 +1,1 @@
-com.teamscale.tia.client.JUnit5TestWiseCoverageExecutionListener
+com.teamscale.tia.client.runlistener.JUnit5TestWiseCoverageExecutionListener
